backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.31
psycopg[binary]==3.1.19
pydantic==2.7.4
python-dotenv==1.0.1